moved query wrapper, fixed useQuery hook to no refresh every second
