+++ pages/_app.tsx
@@ -1,12 +1,6 @@
 import '../styles/globals.css';
-import { QueryClient, QueryClientProvider } from 'react-query';
-const queryClient = new QueryClient();
 
 function MyApp({ Component, pageProps }) {
-  return (
-    <QueryClientProvider client={queryClient}>
-      <Component {...pageProps} />
-    </QueryClientProvider>
-  );
+  return <Component {...pageProps} />;
 }
 export default MyApp;
